Customize the home page
Add the header and the sign-in button inside of it.

Also, display the Add documents Button which in this case as the documents is an empty array, should always display the div with the add document button and create a new room based on the clerk's userId and email.

Also, check if the user is signedIn or not if not redirect them to the sign-in page.