feat: Add defaultAccesses to createDocument action
This commit adds the `defaultAccesses` property to the `createDocument` action in the `room.actions.ts` file. The `defaultAccesses` array is set to `["room:write"]`, allowing users to have write access to the created room by default. This enhancement improves the functionality of the `createDocument` action by providing a default access level for new rooms.